Moved TestingCell to its own file
